requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
psycopg2-binary==2.9.9
django-filter>=24.3
django-cors-headers>=4.5
drf-yasg[validation]>=1.21
